Fixing an issue in TaskExecution's tests related to memory access (optimization(?)).
The tests for TaskExecution and TaskExecution<T> have a number of loops similar to `while (!complete)`, where `complete` is a local Boolean that is set to true from outside the task.

This works just fine in Debug builds, but in Release builds, I assume the JIT does some reachability or loop analysis, sees that `complete` is not set from within the loop, and performs codegen as if it were truly an infinite loop (or possibly it tries to enregister `complete`, and that's never updated with the actual variable's value from memory).

Changing the reads from/writes to `complete` to go through the Volatile functions fixes this since the loop now always definitely checks the current value of the variable.