Adding a primitive configuration capability to IocContainer.
This is not a very highly-capable feature, like ASP.NET's IOptions - it's really just meant as a standard way to provide configuration values.

Also removed the dependency from Ioc to Core to keep Ioc as separated as possible. Core also references UI-adjacent things (ICommand for instance), which isn't needed in non-UI projects that may just want to use IConfiguration.